Expire sessions at approximately the right time.
This prevents us from having a looping thread traversing all the
things regularly, but it doesn't prevent having multiple in-flight
expirerers, so if something with long-lived sessions is flapping, it
could be a bit messy.  There are numerous ways to handle that
situation better that I may care about someday.