Log the connection source.
This is kind of ugly, but SockAddr doesn't give a convenient way to
just display the hostname without the port and the implemntation
involves unsafePerformIO, so I'll just do the easy thing.

Also, it's not obvious how to get the peer address from a WS
connection, so that doesn't work currently.